requirements.txt
asgiref==3.4.1
coverage==5.5
Django==3.2.5
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2==2.9.1
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.8
django-heroku==0.3.1
gunicorn==20.1.0